refactor: move the bearer_request test into auth::raw
