+++ src/auth.rs
@@ -660,19 +660,3 @@ pub async fn verify_tokens(token: &Token) -> Result<Response<crate::user::Twitte
     let req = get(links::auth::VERIFY_CREDENTIALS, token, None);
     request_with_json_response(req).await
 }
-
-#[cfg(test)]
-mod tests {
-    use super::bearer_request;
-
-    #[test]
-    fn bearer_header() {
-        let con_key = "xvz1evFS4wEEPTGEFPHBog";
-        let con_secret = "L8qq9PZyRg6ieKGEKhZolGC0vJWLw8iEJ88DRdyOg";
-        let con_token = super::KeyPair::new(con_key, con_secret);
-
-        let output = bearer_request(&con_token);
-
-        assert_eq!(output, "Basic eHZ6MWV2RlM0d0VFUFRHRUZQSEJvZzpMOHFxOVBaeVJnNmllS0dFS2hab2xHQzB2SldMdzhpRUo4OERSZHlPZw==");
-    }
-}
